jjg7

+++ app/src/main/java/com/ezen/lolketing/adapter/CommentAdapter.java
@@ -4,11 +4,9 @@ import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.view.LayoutInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.PopupMenu;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -19,6 +17,7 @@ import com.ezen.lolketing.R;
 import com.ezen.lolketing.model.BoardDTO;
 import com.firebase.ui.firestore.FirestoreRecyclerAdapter;
 import com.firebase.ui.firestore.FirestoreRecyclerOptions;
+import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.FieldValue;
 import com.google.firebase.firestore.FirebaseFirestore;
 
@@ -27,6 +26,7 @@ import java.text.SimpleDateFormat;
 public class CommentAdapter extends FirestoreRecyclerAdapter<BoardDTO.commentDTO, CommentAdapter.CommentHolder> {
 
     private String documentID;
+    FirebaseAuth auth = FirebaseAuth.getInstance();
 
     public CommentAdapter(@NonNull FirestoreRecyclerOptions<BoardDTO.commentDTO> options, String documentID) {
         super(options);
@@ -34,42 +34,24 @@ public class CommentAdapter extends FirestoreRecyclerAdapter<BoardDTO.commentDTO
     }
 
     @Override
-    protected void onBindViewHolder(@NonNull final CommentHolder holder, int position, @NonNull BoardDTO.commentDTO comment) {
+    protected void onBindViewHolder(@NonNull final CommentHolder holder, int position, @NonNull final BoardDTO.commentDTO comment) {
         SimpleDateFormat timeFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm");
         String timestamp = timeFormat.format(comment.getTimestamp());
 
         holder.textView_userId.setText(comment.getUserId());
         holder.textView_timestamp.setText(timestamp);
         holder.textView_comment.setText(comment.getComment());
-        // 더보기 메뉴 클릭 -> 수정, 삭제 선택 드랍메뉴
-        holder.imageView_more.setOnClickListener(new View.OnClickListener() {
+        // 댓글 삭제 버튼 클릭 이벤트
+        holder.imageView_delete.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                //Creating the instance of PopupMenu
-                PopupMenu popup = new PopupMenu(holder.itemView.getContext(), holder.imageView_more);
-                //Inflating the Popup using xml file
-                popup.getMenuInflater().inflate(R.menu.menu_comment, popup.getMenu());
-                popup.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
-                    public boolean onMenuItemClick(MenuItem item) {
-                        switch (item.getItemId()) {
-                            case R.id.modify:
-                                Toast.makeText(holder.itemView.getContext(),"댓글 수정", Toast.LENGTH_SHORT).show();
-                                return true;
-                            case R.id.delete:
-                                Toast.makeText(holder.itemView.getContext(),"댓글 삭제", Toast.LENGTH_SHORT).show();
-                                CommentDelete(holder.itemView.getContext(), holder.getAdapterPosition());
-                                return true;
-                        }
-                        return false;
-                    }
-                });
-                popup.show();
+                CommentDelete(holder.itemView.getContext(), holder.getAdapterPosition(), comment.getEmail());
              }
         });
 
     } // onBindViewHolder
 
-    public void CommentDelete(final Context context, final int position) {
+    public void CommentDelete(final Context context, final int position, final String email) {
         AlertDialog.Builder builder = new AlertDialog.Builder(context);
 
         builder.setTitle("댓글 삭제").setMessage("정말 삭제하시겠습니까?");
@@ -77,9 +59,13 @@ public class CommentAdapter extends FirestoreRecyclerAdapter<BoardDTO.commentDTO
             @Override
             public void onClick(DialogInterface dialog, int id)
             {
-                getSnapshots().getSnapshot(position).getReference().delete();
-                Toast.makeText(context, "댓글이 삭제 되었습니다.", Toast.LENGTH_SHORT).show();
-                FirebaseFirestore.getInstance().collection("Board").document(documentID).update("commentCounts", FieldValue.increment(-1));
+                if(email.equals(auth.getCurrentUser().getEmail())){
+                    getSnapshots().getSnapshot(position).getReference().delete();
+                    Toast.makeText(context, "댓글이 삭제 되었습니다.", Toast.LENGTH_SHORT).show();
+                    FirebaseFirestore.getInstance().collection("Board").document(documentID).update("commentCounts", FieldValue.increment(-1));
+                }else{
+                    Toast.makeText(context, "삭제 권한이 없습니다", Toast.LENGTH_SHORT).show();
+                }
 //                notifyDataSetChanged();
             }
         });
@@ -104,14 +90,14 @@ public class CommentAdapter extends FirestoreRecyclerAdapter<BoardDTO.commentDTO
 
     public class CommentHolder extends RecyclerView.ViewHolder {
         TextView textView_userId, textView_timestamp, textView_comment;
-        ImageView imageView_more;
+        ImageView imageView_delete;
 
         public CommentHolder(@NonNull View itemView) {
             super(itemView);
             textView_userId = itemView.findViewById(R.id.textView_userId);
             textView_timestamp = itemView.findViewById(R.id.textView_timestamp);
             textView_comment = itemView.findViewById(R.id.textView_comment);
-            imageView_more = itemView.findViewById(R.id.imageView_more);
+            imageView_delete = itemView.findViewById(R.id.imageView_delete);
         }
     } // CommentHolder
